fix password show button
