+++ frontend/app/private/authentication/register/[client]/page.tsx
@@ -1,0 +1,7 @@
+import styles from '@/app/private/authentication/authentication.module.css';
+
+export default function SignupForm() {
+  return (
+    <div></div>
+  );
+}
